test: add git pull integrated testing
